LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/onpaint-css-animation-expected.txt


FAIL onpaint should fire for ticking CSS animations assert_equals: onpaint should fire as the css animation ticks. expected 1 but got 0
